py-embed/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
torch==2.2.0
ftfy==6.1.3
regex==2023.12.25
tqdm==4.66.1
git+https://github.com/openai/CLIP.git
Pillow==10.2.0
python-multipart==0.0.9
pydantic==2.6.1
numpy==1.26.4 